This fixes all lint issues (except for false positives) in:
cc/base/
cc/debug/
cc/resources/
cc/quads/

BUG=144577

Review URL: https://codereview.chromium.org/13051003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@190715 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: adb08548d49107d17ccd2caea44c7fc8bc106680

+++ cc/debug/overdraw_metrics.h
@@ -21,7 +21,7 @@ class OverdrawMetrics {
  public:
   static scoped_ptr<OverdrawMetrics> Create(bool record_metrics_for_frame) {
     return make_scoped_ptr(new OverdrawMetrics(record_metrics_for_frame));
- }
+  }
 
   // These methods are used for saving metrics during update/commit.
 
@@ -93,9 +93,9 @@ class OverdrawMetrics {
   // Count of tiles that were invalidated but not uploaded.
   int tiles_culled_for_upload_;
   // Count the number of bytes in contents textures.
-  unsigned long long contents_texture_use_bytes_;
+  uint64 contents_texture_use_bytes_;
   // Count the number of bytes in RenderSurfaceImpl textures.
-  unsigned long long render_surface_texture_use_bytes_;
+  uint64 render_surface_texture_use_bytes_;
 
   // These values are used for saving metrics during draw.
